Fixes #8949. Promote camel-quarkus-mcp-server to native mode support
The engine needed no changes to work in native mode: the native image
builds without any reflection or resource registration and all engine
conformance scenarios pass against the native binary.

- Move the extension to extensions/ and the tests to integration-tests/,
  drop the JVM only warning build step and declare nativeSince
- Add a native profile and McpServerIT to the test module
- Assign mcp-server to the group-12 test category
- Use McpAssured from quarkus-mcp-server-test to drive the server instead
  of the MCP Java SDK client, which removes the mcp-core and
  mcp-json-jackson2 test dependencies together with their version
  property, test BOM and dependabot entries
- Regenerate the extension metadata, documentation and flattened BOMs
