+++ docs/modules/ROOT/examples/others/mcp-server.yml
@@ -2,11 +2,11 @@
 # This file was generated by camel-quarkus-maven-plugin:update-extension-doc-page
 cqArtifactId: camel-quarkus-mcp-server
 cqArtifactIdBase: mcp-server
-cqNativeSupported: false
-cqStatus: Preview
+cqNativeSupported: true
+cqStatus: Stable
 cqDeprecated: false
 cqJvmSince: 3.39.0
-cqNativeSince: n/a
+cqNativeSince: 3.39.0
 cqCamelPartName: mcp-server
 cqCamelPartTitle: MCP Server
 cqCamelPartDescription: Expose ai-tool routes as MCP tools over streamable HTTP
