+++ integration-tests-jvm/pom.xml
@@ -75,7 +75,6 @@
         <module>ldif</module>
         <module>lucene</module>
         <module>main-devmode</module>
-        <module>mcp-server</module>
         <module>mvel</module>
         <module>opensearch</module>
         <module>printer</module>
