+++ integration-tests/mcp-server/src/main/resources/application.properties
