+++ pom.xml
@@ -190,7 +190,6 @@
         <ftpserver.version>${ftpserver-version}</ftpserver.version>
         <hamcrest.version>3.0</hamcrest.version><!-- Awaitility and Wiremock -->
         <htmlunit-driver.version>4.13.0</htmlunit-driver.version>
-        <mcp-java-sdk.version>2.0.0</mcp-java-sdk.version><!-- https://repo1.maven.org/maven2/io/modelcontextprotocol/sdk/mcp-core/ -->
         <microsoft.graph.version>6.67.0</microsoft.graph.version>
         <microsoft.graph.auth.version>0.3.0</microsoft.graph.auth.version>
         <pdfbox.version>${pdfbox-version}</pdfbox.version>
