+++ tooling/scripts/test-categories.yaml
@@ -239,6 +239,7 @@ group-12:
   - knative
   - langchain4j-web-search
   - langchain4j-tools
+  - mcp-server
   - openapi-java
   - openapi-validator
   - paho-mqtt5
